+++ fuocore/core/provider.py
@@ -1,10 +1,28 @@
 # -*- coding: utf-8 -*-
 
 from abc import ABC, abstractmethod
+from fuocore.models import (
+    SongModel,
+    ArtistModel,
+    AlbumModel,
+    PlaylistModel,
+    LyricModel,
+
+    UserModel,
+)
 
 
 class AbstractProvider(ABC):
-    """abstract music resource provider"""
+    """abstract music resource provider
+    """
+
+    # A well behaved provider should implement its own models .
+    Song = SongModel
+    Artist = ArtistModel
+    Album = AlbumModel
+    Playlist = PlaylistModel
+    Lyric = LyricModel
+    User = UserModel
 
     @property
     @abstractmethod