+++ fuocore/daemon/main.py
@@ -1,19 +1,66 @@
 import asyncio
 import logging
 import sys
+from collections import defaultdict
+from urllib.parse import urlparse
 
 from .aio_tcp_server import TcpServer
 from .pubsub import run as run_pubsub
 
 from fuocore import setup_logger
-from fuocore import MpvPlayer, Source
+from fuocore import MpvPlayer
 from fuocore import LiveLyric
-from fuocore.core.provider import providers
+from fuocore.core.furi import parse_furi
 from fuocore.protocol.parser import CmdParser
 from fuocore.protocol.handlers import exec_cmd
-from fuocore.plugin import load_plugins
+from fuocore.local.provider import provider as local_provider
+from fuocore.netease.provider import provider as netease_provider
 
 logger = logging.getLogger(__name__)
+
+
+class Source(object):
+    def __init__(self, prvs=[]):
+        self.providers = set(prvs)
+
+    def register(self, provider):
+        self.providers.add(provider)
+
+    def search(self, keyword):
+        """search song by keyword"""
+        songs = []
+        for provider in self.providers:
+            result = provider.search(keyword=keyword)
+            _songs = list(result.songs[:20])
+            logger.debug('在 provider({}) 中搜索到了 {} 首歌曲'
+                         .format(provider.name, len(_songs)))
+            songs.extend(_songs)
+        logger.debug('共搜索到了 {} 首关于 {} 的歌曲'.format(len(songs), keyword))
+        return songs
+
+    def _get_provider(self, name):
+        for provider in self.providers:
+            if provider.identifier == name:
+                return provider
+        return None
+
+    def get_song(self, furi_str):
+        """get song from identifier"""
+        result = urlparse(furi_str)
+        provider = self._get_provider(result.netloc)
+        identifier = result.path.split('/')[-1]
+        return provider.Song.get(identifier)
+
+    def list_songs(self, furi_str_list):
+        provider_songs_map = defaultdict(list)
+        for furi_str in furi_str_list:
+            furi = parse_furi(furi_str)
+            provider_songs_map[furi.provider].append(furi.identifier)
+        songs = []
+        for provider_name, identifiers in provider_songs_map.items():
+            provider = self._get_provider(provider_name)
+            songs += provider.Song.list(identifiers)
+        return songs
 
 
 class App(object):
@@ -24,15 +71,16 @@ class App(object):
 
     def initialize(self):
         self.player.initialize()
-        load_plugins()
+        self.source.register(local_provider)
+        self.source.register(netease_provider)
 
     def list_providers(self):
-        for provider in providers:
+        for provider in self.source.providers:
             yield provider
 
     def get_provider(self, name):
-        for provider in providers:
-            if provider.name == name:
+        for provider in self.source.providers:
+            if provider.identifier == name:
                 return provider
         return None